+++ daml-lf/tests/scenario/actualize.sh
@@ -18,11 +18,10 @@ TESTMAIN=$1
 TESTDIR="$(dirname $TESTMAIN)"
 TESTDAR="$TESTDIR/Main.dar"
 BAZEL_BIN="$(bazel info bazel-bin)"
-GHC_PRIM_DALF=${BAZEL_BIN}/daml-foundations/daml-ghc/package-database/deprecated/daml-prim-1.3.dalf
 REGEX_HIDE_HASHES="s,@[a-z0-9]{8},@XXXXXXXX,g"
 
 bazel build //daml-foundations/daml-tools/da-hs-damlc-app:da-hs-damlc-app
-../../../bazel-bin/daml-foundations/daml-tools/da-hs-damlc-app/da-hs-damlc-app package --debug --target 1.3 $TESTMAIN main -o $TESTDAR
+../../../bazel-bin/daml-foundations/daml-tools/da-hs-damlc-app/da-hs-damlc-app package --debug --target 1.dev $TESTMAIN main -o $TESTDAR
 
 bazel build //daml-lf/repl:repl
 ../../../bazel-bin/daml-lf/repl/repl test Test:run $TESTDAR | sed '1d' | sed -E "$REGEX_HIDE_HASHES" | tee ${TESTDIR}/EXPECTED.ledger.new