+++ docs/source/support/release-notes.rst
@@ -25,6 +25,13 @@ Navigator
 - Fixed a regression where Navigator console was not able to inspect contracts and events.
   See `#1454 <https://github.com/digital-asset/daml/issues/1454>`__.
 
+
+Sandbox
+~~~~~~~
+
+- Added recovery around failing ledger entry persistence queries using Postgres. See `#1505 <https://github.com/digital-asset/daml/pull/1505>`__.
+
+
 0.12.22 - 2019-05-29
 --------------------