+++ docs/source/support/release-notes.rst
@@ -14,6 +14,9 @@ DAML Compiler
 
 - **BREAKING CHANGE**: Drop support for DAML-LF 1.3. Compiling to DAML-LF 1.4 should work without any code changes, although we highly recommend not specifying a target DAML-LF version at all. (The ledger server still supports DAML-LF 1.3.)
 
+- Fix initialization of package-db for non-default DAML-LF versions.
+  This fixes issues when using "daml build --target 1.3" (or other target versions).
+
 DAML Standard Library
 ~~~~~~~~~~~~~~~~~~~~~