+++ docs/source/support/release-notes.rst
@@ -70,7 +70,7 @@ Sandbox
 ~~~~~~~
 
 - Fixed a bug in the SQL backend that caused transactions with a fetch node referencing a contract created in the same transaction to be rejected.
-  See [issue #1435](https://github.com/digital-asset/daml/issues/1435).
+  See `issue #1435 <https://github.com/digital-asset/daml/issues/1435>`__.
 
 0.12.21 - 2019-05-28
 --------------------